# demo-app/openapi.yaml
openapi: 3.0.3
info:
  title: ag-qrew Demo Task Manager API
  version: 1.0.0
  description: >
    REST API for the demo task manager. This spec documents the CORRECT,
    intended behaviour. The running app deliberately deviates in 4 places
    (see docs/PLANTED_BUGS.md) — those deviations are what the QA agents detect.
servers:
  - url: http://localhost:3000
tags:
  - name: auth
    description: Authentication
  - name: tasks
    description: Task CRUD
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@demo.test
        password:
          type: string
          example: admin123
    LoginResponse:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
          description: Non-empty bearer token (base64 of email).
        role:
          type: string
          enum: [admin, standard]
    Task:
      type: object
      required: [id, title, done, ownerRole]
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          maxLength: 200
        done:
          type: boolean
        ownerRole:
          type: string
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
          description: Required. Must be 1..200 characters.
        done:
          type: boolean
          default: false
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        done:
          type: boolean
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
paths:
  /api/auth/login:
    post:
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks:
    get:
      tags: [tasks]
      summary: List all tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [tasks]
      summary: Create a task
      description: >
        Creates a task. `title` is required and limited to 200 characters.
        A missing/empty title MUST return 400. An over-length title (>200)
        MUST return 400. On success returns 201 with the created task.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Missing/empty title or title longer than 200 characters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks/{id}:
    put:
      tags: [tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [tasks]
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: boolean
                    example: true
        '401':
          description: Missing or invalid Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'